# api-gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LearnBot API Gateway
  description: |
    The LearnBot API Gateway provides a unified REST API for the AI-powered career
    development platform. It handles authentication, user profiles, job matching,
    skill gap analysis, and personalized training recommendations.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained via `/api/auth/login` or
    `/api/auth/register`. Include it in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    The API enforces a rate limit of 10 requests/second per IP with a burst of 30.
    Exceeding the limit returns `429 Too Many Requests`.

    ## Response Format
    All responses follow a consistent envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "total": 100, "limit": 20, "offset": 0 }
    }
    ```
    Error responses:
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "request validation failed",
        "details": [{ "field": "email", "message": "must be a valid email" }]
      }
    }
    ```
  version: 1.0.0
  contact:
    name: LearnBot Team
    url: https://github.com/bramrahmadi/LearnBot
servers:
  - url: http://localhost:8090
    description: Local development server
...
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: Creates a new user account and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "jane@example.com"
              password: "securepassword123"
              full_name: "Jane Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "jane@example.com"
              password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ─────────────────────────────────────────────────────────────────────────────
  # Profile
  # ─────────────────────────────────────────────────────────────────────────────
...
  /api/analysis/gaps:
    post:
      tags: [Analysis]
      summary: Analyze skill gaps for a target job
      description: |
        Compares the current user's skills against a job's requirements and
        identifies missing skills categorized as critical, important, or nice-to-have.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GapAnalysisRequest'
            examples:
              byJobID:
                summary: Analyze gaps for a specific job
                value:
                  job_id: "job-001"
              inlineJob:
                summary: Analyze gaps for an inline job description
                value:
                  job:
                    title: "Senior Go Engineer"
                    required_skills: ["Go", "PostgreSQL", "Docker", "Kubernetes"]
                    preferred_skills: ["Terraform", "AWS"]
                    min_years_experience: 5
                    experience_level: "senior"
      responses:
        '200':
          description: Gap analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GapAnalysisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ─────────────────────────────────────────────────────────────────────────────
  # Training
  # ─────────────────────────────────────────────────────────────────────────────
  /api/training/recommendations:
    get:
      tags: [Training]
      summary: Get personalized training plan (GET)
      description: Generates a learning plan for the current user targeting a specific job.
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: query
          schema:
            type: string
          example: "job-001"
      responses:
        '200':
          description: Training plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingPlanResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Training]
      summary: Get personalized training plan (POST with preferences)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingRecommendationRequest'
            example:
              job_id: "job-001"
              preferences:
                prefer_free: false
                max_budget_usd: 100
                weekly_hours_available: 10
                prefer_hands_on: true
                prefer_certificates: false
      responses:
        '200':
          description: Training plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingPlanResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ─────────────────────────────────────────────────────────────────────────────
  # Resources
  # ─────────────────────────────────────────────────────────────────────────────
  /api/resources/search:
    get:
      tags: [Resources]
      summary: Search learning resources
      description: |
        Searches the built-in catalog of 60+ curated learning resources.
        No authentication required.
      parameters:
        - name: skill
          in: query
          description: Filter by skill name (e.g. "Python", "Go")
          schema:
            type: string
        - name: type
          in: query
          description: Filter by resource type
          schema:
            type: string
            enum: [course, certification, documentation, video, book, practice, article, project]
        - name: difficulty
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [beginner, intermediate, advanced, expert, all_levels]
        - name: free
          in: query
          description: Return only free resources
          schema:
            type: boolean
        - name: has_certificate
          in: query
          description: Return only resources with certificates
          schema:
            type: boolean
        - name: has_hands_on
          in: query
          description: Return only resources with hands-on exercises
          schema:
            type: boolean
        - name: min_rating
          in: query
          description: Minimum rating (0.0-5.0)
          schema:
            type: number
            minimum: 0
            maximum: 5
        - name: q
          in: query
          description: Full-text search query
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results (default 20, max 100)
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Resources retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceListResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
          example: "jane@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
        full_name:
          type: string
          example: "Jane Doe"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    GapAnalysisRequest:
      type: object
      properties:
        job_id:
          type: string
          description: ID of a job in the catalog
        job:
          $ref: '#/components/schemas/JobRequirementsInput'
    JobRequirementsInput:
      type: object
      required: [required_skills]
      properties:
        title:
          type: string
        required_skills:
          type: array
          items:
            type: string
        preferred_skills:
          type: array
          items:
            type: string
        min_years_experience:
          type: number
        required_degree_level:
          type: string
        location_type:
          type: string
        industry:
          type: string
        experience_level:
          type: string
    TrainingRecommendationRequest:
      type: object
      properties:
        job_id:
          type: string
        job:
          $ref: '#/components/schemas/JobRequirementsInput'
        preferences:
          $ref: '#/components/schemas/LearningPreferences'
    LearningPreferences:
      type: object
      properties:
        prefer_free:
          type: boolean
          default: false
        max_budget_usd:
          type: number
          minimum: 0
        weekly_hours_available:
          type: number
          default: 10
        prefer_hands_on:
          type: boolean
          default: false
        prefer_certificates:
          type: boolean
          default: false
        target_date:
          type: string
          format: date
          example: "2025-09-01"
        preferred_resource_types:
          type: array
          items:
            type: string
            enum: [course, certification, documentation, video, book, practice]
        excluded_providers:
          type: array
          items:
            type: string

    # ─── Response schemas ──────────────────────────────────────────────────────
    AuthSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              description: JWT access token
            expires_at:
              type: string
              format: date-time
            user:
              $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
...
    GapAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            critical_gaps:
              type: array
              description: Must-have skills the candidate is missing
            important_gaps:
              type: array
              description: Preferred skills the candidate is missing
            nice_to_have_gaps:
              type: array
            total_gaps:
              type: integer
            critical_gap_count:
              type: integer
            important_gap_count:
              type: integer
            readiness_score:
              type: number
              description: Overall readiness score [0-100]
            total_estimated_learning_hours:
              type: integer
            top_priority_gaps:
              type: array
            matched_skills:
              type: array
              items:
                type: string
            visual_data:
              type: object
    TrainingPlanResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            job_title:
              type: string
            readiness_score:
              type: number
            total_gaps:
              type: integer
            total_estimated_hours:
              type: number
            phases:
              type: array
              description: Ordered learning phases (Critical → Preferred → Nice-to-have)
            timeline:
              type: object
              description: Week-by-week study schedule
            matched_skills:
              type: array
              items:
                type: string
            summary:
              type: object
              properties:
                headline:
                  type: string
                  example: "📚 3 critical gaps to close – estimated 8 weeks"
                critical_gap_count:
                  type: integer
                free_resource_count:
                  type: integer
                paid_resource_count:
                  type: integer
                estimated_total_cost_usd:
                  type: number
                top_skills_to_learn:
                  type: array
                  items:
                    type: string
                quick_wins:
                  type: array
                  items:
                    type: string
    ResourceListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ResourceEntry'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResourceEntry:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
        provider:
          type: string
        resource_type:
          type: string
          enum: [course, certification, documentation, video, book, practice, article, project]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert, all_levels]
        cost_type:
          type: string
          enum: [free, freemium, paid, subscription, free_audit]
        cost_usd:
          type: number
        duration_hours:
          type: number
        duration_label:
          type: string
        skills:
          type: array
          items:
            type: string
        primary_skill:
          type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        rating_count:
          type: integer
        has_certificate:
          type: boolean
        has_hands_on:
          type: boolean
        is_verified:
          type: boolean
    ResponseMeta:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "request validation failed"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "request validation failed"
              details:
                - field: "email"
                  message: "must be a valid email address"
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "missing or invalid Authorization header"
...
    ConflictError:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "EMAIL_TAKEN"
              message: "an account with this email already exists"